+++ scripts/build.ps1
@@ -6,12 +6,14 @@ $gitBranch = $env:APPVEYOR_REPO_BRANCH
 $extraParameters = if ($gitBranch -eq "master") {
     "-DHALORAY_VERSION='${env:APPVEYOR_BUILD_VERSION}'"
 } else {
-    ""
+    "-DGIT_BRANCH='${gitBranch}'"
 }
 
 pushd
 mkdir build
 cd build
+
+pushd
 cmake .. -G "Visual Studio 15 2017 Win64" $extraParameters
 cmake --build . --config Release
 cd .\src\Release\
@@ -23,7 +25,22 @@ cd .\src\Release\
     --no-webkit2 `
     --no-angle `
     --no-opengl-sw `
-    -no-system-d3d-compiler `
+    --no-system-d3d-compiler `
     --release `
     haloray.exe
+popd
+
+pushd
+cd .\tests\Release\
+& "${env:Qt5_DIR}\bin\windeployqt.exe" `
+    --no-quick-import `
+    --no-translations `
+    --no-webkit2 `
+    --no-angle `
+    --no-opengl-sw `
+    --no-system-d3d-compiler `
+    --release `
+    .
+popd
+
 popd