+++ CHANGELOG.md
@@ -5,10 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.4.0 - 2019-07-29
+
+### Added
+- Show error dialog and exit if GPU does not support OpenGL 4.4
+
+### Fixed
+- Bug where changing multiple scattering probability did not trigger a new
+  render if maximum iterations were reached
+- Crashing on Intel GPUs with Linux and Mesa3D
+
 ## 2.3.1 - 2019-07-14
 
 ### Fixed
-- Fixed build configuration bug where resources were not being linked on Linux
+- Build configuration bug where resources were not being linked on Linux
 
 ## 2.3.0 - 2019-07-13